move help to lib
