+++ __tests__/application.test.ts
@@ -1,6 +1,6 @@
 import { createApplication } from "@/app/actions/actions";
 import { validatePlanningParams } from "@/app/actions/validator";
-import { verifyApiKey } from "../util/helpers/apiKey";
+import { verifyApiKey } from "../lib/apiKey";
 import handler from "@/app/api/application/route";
 
 jest.mock("../util/actions/actions");
