+++ lib/type.ts
