+++ src/app/planning-applications/[id]/page.tsx
@@ -4,7 +4,7 @@ import Breadcrumbs from "@/components/breadcrumbs";
 import About from "@/components/about";
 import Impact from "@/components/impact";
 import Process from "@/components/process";
-import { DataDetails } from "../../../../util/helpers/type";
+import { DataDetails } from "../../../../lib/type";
 import { getApplicationById } from "@/app/actions/actions";
 import moment from "moment";
 import { useParams } from "next/navigation";
