+++ src/components/planning-application/index.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import Image from "next/image";
 import { LocalIcon } from "../../../public/assets/icons";
 import { urlFor } from "@/app/actions/client";
-import { Data } from "../../../util/helpers/type";
+import { Data } from "../../../lib/type";
 import { distanceInMiles } from "../../app/actions/actions";
 
 const PlanningApplications = ({
